clean up Marble/timeline component
