+++ components/marble-avatar/colors.ts
@@ -1,0 +1,24 @@
+export const colors = [
+  '#f76b15', // orange-9
+  '#e54d2e', // tomato-9
+  '#e5484d', // red-9
+  '#e54666', // ruby-9
+  '#e93d82', // crimson-9
+  '#d6409f', // pink-9
+  '#ab4aba', // plum-9
+  '#8e4ec6', // purple-9
+  '#6e56cf', // violet-9
+  '#5b5bd6', // iris-9
+  '#3e63dd', // indigo-9
+  '#0090ff', // blue-9
+  '#75C7F0', // sky-11
+  '#00a2c7', // cyan-9
+  '#58D5BA', // mint-11
+  '#12a594', // teal-9
+  '#29a383', // jade-9
+  '#30a46c', // green-9
+  '#46a758', // grass-9
+  '#bdee63', // lime-9
+  '#ffe629', // yellow-9
+  '#ffc53d', // amber-9
+]
